app/requirements.txt
dlib==19.22.1
pymongo==3.8.0
numpy==1.21.4
tqdm==4.32.2
names==0.3.0
scikit_image==0.19.0
imutils==0.5.2
opencv_python==4.5.4.60
Pillow==6.1.0
pgi==0.0.11.2
vext==0.7.6
vext.gi==0.7.4